Update README.md
Hi Super Joomlers,

Its seems that the Content Security Policy rule regarding frame-src doesn't seem to work for me without the https://*.hcaptcha.com  see the extra ``` * ``` at least for Firefox it did not seem to work as expected. But with the * before the domain it works. Just a feedback. Not trying to be the "smart guy". Just a Super Joomler like you.